requirements.txt
flask
flask-sqlalchemy
flask-view-counter
feedparser
python-dateutil